+++ src/lib/auth/AuthProviders/AuthProviders.ts
@@ -1,0 +1,10 @@
+import { AuthProviderBase } from './AuthProviderBase'
+import { OAuthAuthenticator } from './OAuthAuthenticator'
+
+export class AuthProviders {
+  public static getAuthProviders(): Record<string, AuthProviderBase> {
+    return {
+      [OAuthAuthenticator.AUTH_TYPE]: new OAuthAuthenticator(),
+    }
+  }
+}