+++ src/modules/auth/logout.ts
@@ -1,9 +1,9 @@
-import { clear } from '../../conf'
+import { SessionManager } from '../../lib/session/SessionManager'
 import log from '../../logger'
 
 export default () => {
   log.debug('Clearing config file')
-  clear()
+  const sessionManager = SessionManager.getSessionManager()
+  sessionManager.logout()
   log.info('See you soon!')
-  return Promise.resolve()
 }