+++ src/modules/auth/switch.ts
@@ -1,10 +1,9 @@
 import chalk from 'chalk'
 import { split } from 'ramda'
-import { getAccount, getLastUsedAccount, getLogin, getTokens, getWorkspace } from '../../conf'
+import { getAccount, getLastUsedAccount, getLogin, getWorkspace } from '../../conf'
 import { CommandError } from '../../errors'
+import { SessionManager } from '../../lib/session/SessionManager'
 import log from '../../logger'
-import { Token } from '../../Token'
-import loginCmd, { saveCredentials } from './login'
 
 export const switchAccount = async (account: string, options, previousAccount = getAccount()) => {
   const isValidAccount = /^\s*[\w-]+\s*$/.test(account)
@@ -18,19 +17,12 @@ export const switchAccount = async (account: string, options, previousAccount = 
     throw new CommandError(`You're already using the account ${chalk.blue(account)}`)
   }
 
-  const accountToken = new Token(getTokens()[account])
-  if (accountToken.isValid()) {
-    log.debug(`Token stored for ${account}/${accountToken.login} is still valid`)
-    saveCredentials(accountToken.login, account, accountToken.token, workspace)
-    log.info(
-      `Logged into ${chalk.blue(getAccount())} as ${chalk.green(getLogin())} at workspace ${chalk.green(
-        getWorkspace()
-      )}`
-    )
-  } else {
-    log.debug(`Token for ${account} isn't stored or isn't valid`)
-    return loginCmd({ account, workspace })
-  }
+  const sessionManager = SessionManager.getSessionManager()
+  await sessionManager.login(account, { targetWorkspace: workspace })
+
+  log.info(
+    `Logged into ${chalk.blue(getAccount())} as ${chalk.green(getLogin())} at workspace ${chalk.green(getWorkspace())}`
+  )
 }
 
 const hasAccountSwitched = (account: string) => {